post

+++ app/Http/Controllers/TasksController.php
@@ -14,14 +14,22 @@ class TasksController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function index()
-    {
-        $tasks = Task::all();
+        {
+             $data = [];
+             if (\Auth::check()) {
+                 $user = \Auth::user();
+                 $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+                 
+                 $data = [
+                     'user' => $user,
+                     'tasks' => $tasks
+                     ];
+             }
+             
+             return view ('welcome', $data);
+        }
         
-        return view('welcome', [
-            'tasks' => $tasks,
-            ]);
-        
-    }
+    
 
     /**
      * Show the form for creating a new resource.
@@ -49,10 +57,10 @@ class TasksController extends Controller
             'status' => 'required|max:10'
             ]);
         
-        $task = new Task;
-        $task->content = $request->content;
-        $task->status = $request->status;
-        $task->save();
+        $request->user()->tasks()->create([
+            'content' => $request->content,
+            'status' => $request->status
+            ]);
         
         return redirect('/');
     }
@@ -119,10 +127,12 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-        $task = Task::find($id);
-        $task->delete();
+        $task = \App\Task::find($id);
         
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
         return redirect('/');
-        //
+        
     }
 }
